+++ router/CHANGELOG.md
@@ -4,6 +4,24 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.347.1](https://github.com/wundergraph/cosmo/compare/router@0.347.0...router@0.347.1) (2026-09-15)
+
+### Bug Fixes
+
+* enable multi-fetch and scheduling ([#3243](https://github.com/wundergraph/cosmo/issues/3243)) ([ce3eec1](https://github.com/wundergraph/cosmo/commit/ce3eec17ab6ac993cb281989dfe8b6f9a35be391)) (@ysmolski)
+
+# [0.347.0](https://github.com/wundergraph/cosmo/compare/router@0.346.1...router@0.347.0) (2026-09-14)
+
+### Features
+
+* entity caching invalidation ([#3213](https://github.com/wundergraph/cosmo/issues/3213)) ([c6363a3](https://github.com/wundergraph/cosmo/commit/c6363a3d6fde9ee49e30d25d1a55de80496681bd)) (@SkArchon)
+
+## [0.346.1](https://github.com/wundergraph/cosmo/compare/router@0.346.0...router@0.346.1) (2026-09-09)
+
+### Bug Fixes
+
+* **router:** cache responses without public directive ([#3221](https://github.com/wundergraph/cosmo/issues/3221)) ([dbf5c7a](https://github.com/wundergraph/cosmo/commit/dbf5c7abb3645e9bd5881a81c486c4aff0c50b62)) (@fiam)
+
 # [0.346.0](https://github.com/wundergraph/cosmo/compare/router@0.345.0...router@0.346.0) (2026-09-08)
 
 ### Bug Fixes